refact(RegisterForm): break down components
