+++ src/pages/RegisterForm/RegisterForm.js
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useState } from "react";
 import { useHistory, useParams, useLocation } from "react-router-dom";
 
-import { CustomerForm, CarForm, AddServiceButton } from "components";
+import { CustomerForm, CarForm } from "components";
 import {
   NOTIFICATION_DURATION,
   NOTIFICATION_TYPES,
@@ -9,14 +9,10 @@ import {
   CombinedFormsProvider,
 } from "context";
 import { useCombinedForms, useLoader, useNotification } from "hooks";
-import {
-  Button,
-  BUTTON_VARIANTS,
-  Modal,
-  ScreenLoader,
-  Text,
-} from "ui-fragments";
+import { ScreenLoader } from "ui-fragments";
 import { engineAPI } from "utils";
+
+import { CreateServiceModal } from "./CreateServiceModal";
 
 const RegisterForm = () => {
   const {
@@ -103,40 +99,6 @@ const RegisterForm = () => {
         customerCarId={customer_cars?.id}
       />
     </div>
-  );
-};
-
-const CreateServiceModal = ({ handleClose, customerCarId, isOpen }) => {
-  const history = useHistory();
-  return (
-    <Modal isOpen={isOpen} handleClose={handleClose} title="Cliente Adicionado">
-      <div>
-        <div className="my-5">
-          <Text>
-            Você pode adicionar uma ordem de serviço para esse cliente
-          </Text>
-        </div>
-        <div className="mt-4 flex gap-3 justify-end">
-          <Button
-            variant={BUTTON_VARIANTS.GHOST}
-            onClick={handleClose}
-            showVariantIcon={false}
-          >
-            Cancelar
-          </Button>
-          <AddServiceButton
-            customerCarId={customerCarId}
-            onServiceAdd={data =>
-              history.push(`/services/${data.id}`, {
-                redirect: `/customers/${customerCarId}`,
-              })
-            }
-          >
-            <Text>Adicionar Também o Serviço</Text>
-          </AddServiceButton>
-        </div>
-      </div>
-    </Modal>
   );
 };
